Refactor LdapURLSet and ConnectionStrategy pattern.
Make the LdapURLSet responsible for tracking available URLs with
ConnectionStrategy adapting to answer a simple question:

Given an LdapURLSet, what is the next active URL that should be used for an
arbitrary consumer function.

State tracking moves to LdapURLSet and the ConnectionFactory is responsible
for producing and managing exactly one instance of this class; the
ConnectionFactory instance is passed around to provide access to both the
ConnectionConfig and the LdapURLSet to components that need it.

A sequence diagram in the docs directory describes the new connection
initialization process.