Use a singlton to activate inactive URLs.
This moves the thread concerns from the strategies to a singleton.
Instead of spinning up and tearing down excutors as needed, a single thread is always allocated for the purpose of activating URLs.
RetryMetadata now holds a reference to a connection strategy so the LdapURL can be activated.
The execution period of the new activator is no longer configurable.