ConnectionFactory shouldn't extends AutoClosable.
A connection factory is generally a long lived object and having it extend AutoClosable is a confusing bit of the API.
While DefaultConnectionFactory could benefit from that interface it does not need to be closed by default and would not generally be used in try block.
Removed the property override for the DefaultConnectionFactory transport; this property would be problematic for stateful Transports since implementations would be unlikely to invoke close on this object.
Add DualThread implementations for convenience testing.