Passivation should occur before validation.
On check-in passivation should occur before validation as it is often necessary to put the connection in a state so that validation will succeed.
Passivation should also occur when creating new connections for the same reason as above and to ensure that periodic validation is always operating on passivated connections.