Refactor merge to support operation config.
The underlying operations used by the merge operation can now be configured with handlers.
Update MergeOperation to expose properties for these operations.
Note that this is a behavioral change to the merge as the initial search is no longer guaranteed to occur on the same connection as the following operation.
Add copy methods to each operation for consistency.