Add validator to SingleConnectionFactory.
The SyncReplRunner attempted to maintain a valid connection using exception handling.
This proved to be difficult to configure and complex to implement.
Using an external thread to ensure the connection is valid has proven to be a simplier and more reliable solution.

Add validator, onOpen and onClose properties to SingleConnectionFactory.
Add onSuccess and onFailure properties to AbstractConnectionValidator.
Update SingleConnectionFactory to execute the validator when it is configured.
Add a ReinitializeConnectionConsumer that SyncReplRunner can leverage to maintain its connection.
Update SyncReplRunner to use a ConnectionValidator instead of exception handling.